fix warning used but marked unused

+++ src/Aruco/ArucoCore.cpp
@@ -7,6 +7,9 @@
 #include "opencv2/highgui/highgui.hpp"
 #include "opencv2/calib3d/calib3d.hpp"
 
+#include <leathers/push>
+#include <leathers/used-but-marked-unused>
+
 namespace ArucoModul {
 
 ArucoCore::ArucoCore()
@@ -46,10 +49,10 @@ const QMatrix4x4 ArucoCore::getDetectedMatrix( int markerId, cv::Mat inputImage 
 	//this->detectMarkers();
 	this->getMatrix( markerId, modelViewMatrix );
 
-	QMatrix4x4 matrix( modelViewMatrix[ 0], modelViewMatrix[ 1], modelViewMatrix[ 2], modelViewMatrix[ 3],
-					   modelViewMatrix[ 4], modelViewMatrix[ 5], modelViewMatrix[ 6], modelViewMatrix[ 7],
-					   modelViewMatrix[ 8], modelViewMatrix[ 9], modelViewMatrix[10], modelViewMatrix[11],
-					   modelViewMatrix[12], modelViewMatrix[13], modelViewMatrix[14], modelViewMatrix[15] );
+	QMatrix4x4 matrix( static_cast<float>( modelViewMatrix[ 0] ), static_cast<float>( modelViewMatrix[ 1] ), static_cast<float>( modelViewMatrix[ 2] ), static_cast<float>( modelViewMatrix[ 3] ),
+					   static_cast<float>( modelViewMatrix[ 4] ), static_cast<float>( modelViewMatrix[ 5] ), static_cast<float>( modelViewMatrix[ 6] ), static_cast<float>( modelViewMatrix[ 7] ),
+					   static_cast<float>( modelViewMatrix[ 8] ), static_cast<float>( modelViewMatrix[ 9] ), static_cast<float>( modelViewMatrix[10] ), static_cast<float>( modelViewMatrix[11] ),
+					   static_cast<float>( modelViewMatrix[12] ), static_cast<float>( modelViewMatrix[13] ), static_cast<float>( modelViewMatrix[14] ), static_cast<float>( modelViewMatrix[15] ) );
 
 	return matrix;
 }
@@ -63,10 +66,10 @@ const QMatrix4x4 ArucoCore::getProjectionMatrix( cv::Mat inputImage )
 	//get projection matrix via ArUco
 	mCamParam.glGetProjectionMatrix( inputImage.size(),inputImage.size(),projectionMatrix,0.01,10000.0 );
 
-	QMatrix4x4 matrix( projectionMatrix[ 0], projectionMatrix[ 1], projectionMatrix[ 2], projectionMatrix[ 3],
-					   projectionMatrix[ 4], projectionMatrix[ 5], projectionMatrix[ 6], projectionMatrix[ 7],
-					   projectionMatrix[ 8], projectionMatrix[ 9], projectionMatrix[10], projectionMatrix[11],
-					   projectionMatrix[12], projectionMatrix[13], projectionMatrix[14], projectionMatrix[15] );
+	QMatrix4x4 matrix( static_cast<float>( projectionMatrix[ 0] ), static_cast<float>( projectionMatrix[ 1] ), static_cast<float>( projectionMatrix[ 2] ), static_cast<float>( projectionMatrix[ 3] ),
+					   static_cast<float>( projectionMatrix[ 4] ), static_cast<float>( projectionMatrix[ 5] ), static_cast<float>( projectionMatrix[ 6] ), static_cast<float>( projectionMatrix[ 7] ),
+					   static_cast<float>( projectionMatrix[ 8] ), static_cast<float>( projectionMatrix[ 9] ), static_cast<float>( projectionMatrix[10] ), static_cast<float>( projectionMatrix[11] ),
+					   static_cast<float>( projectionMatrix[12] ), static_cast<float>( projectionMatrix[13] ), static_cast<float>( projectionMatrix[14] ), static_cast<float>( projectionMatrix[15] ) );
 	return matrix;
 }
 
@@ -151,7 +154,7 @@ void ArucoCore::detectMarkers()
 int ArucoCore::getMatrix( int markerId, double* modelviewmatrix )
 {
 	if ( mMarkers.size() > 0 && markerId != -1 ) {
-		mMarkers[markerId].glGetModelViewMatrix( modelviewmatrix );
+		mMarkers[static_cast<unsigned long>(markerId)].glGetModelViewMatrix( modelviewmatrix );
 		return 0;
 	}
 	return -1;
@@ -262,3 +265,5 @@ void ArucoCore::setBaseMarkerIndex( int baseMarkerIndex )
 }
 
 } // namespace ArucoModul
+
+#include <leathers/pop>
