+++ src/Leap/HandModule/Manipulator/HandMapper.cpp
@@ -98,8 +98,8 @@ Leap::Vector Leap::HandMapper::recalculateDepthNode( Leap::Vector vector, float 
 			else {
 				this->calibrationCounter = 0;
 			}
-			this->calibrationLastPoint.x = this->nodeScreenCoords.x();
-			this->calibrationLastPoint.y = this->nodeScreenCoords.y();
+			this->calibrationLastPoint.x = static_cast<int>( this->nodeScreenCoords.x() );
+			this->calibrationLastPoint.y = static_cast<int>( this->nodeScreenCoords.y() );
 		}
 
 		// keep calibrationCounter from overflow
